Expose a constant when passing a model reference
To allow Packwerk to detect dependency violations, we need to make constant references more explicit. One way to do that is to update the fixturies library to take a constant model instead of a symbol. That way, Packwerk can pick up on the constant and report dependency violations without being updated. This circumstance means that we get dependency enforcement without making significant changes to our codebase.